[refactor] Renamed file-view to file-list.
